# api/ytdownloader.openapi.yaml
openapi: 3.0.2
info:
  title: FastAPI YouTube Downloader
  version: 1.0.0
  description: This is a FastAPI application that allows you to download YouTube videos and retrieve video information.
paths:
  /download/{resolution}:
    post:
      summary: Download Video by Resolution
      description: Downloads a YouTube video with the specified resolution.
      parameters:
        - in: path
          name: resolution
          schema:
            type: string
          required: true
          description: The resolution of the video to be downloaded.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: The YouTube URL of the video to download.
              required:
                - url
      responses:
        '200':
          description: Video downloaded successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Video with resolution {resolution} downloaded successfully.
        '400':
          description: Invalid YouTube URL or missing URL parameter.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Invalid YouTube URL.
        '500':
          description: Error during video download.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: An error occurred.
  /video_info:
    post:
      summary: Get Video Information
      description: Retrieves information about a YouTube video.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: The YouTube URL of the video.
              required:
                - url
      responses:
        '200':
          description: Video information retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: string
                    example: Example Video Title
                  author:
                    type: string
                    example: Example Author
                  length:
                    type: integer
                    example: 120
                  views:
                    type: integer
                    example: 1000
                  description:
                    type: string
                    example: This is an example video description.
                  publish_date:
                    type: string
                    format: date
                    example: 2020-01-01
        '400':
          description: Invalid YouTube URL or missing URL parameter.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: Invalid YouTube URL.
        '500':
          description: Error retrieving video information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: An error occurred.
components:
  schemas:
    VideoURL:
      type: object
      properties:
        url:
          type: string
      required:
        - url